prepped the path notes MD file for when I get finished with the Simple Items update;
I've updated the README.md so show the last STABLE version of the program.

With the last stable version instead of the latest actual version, I won't have to make sure to update the README every time I make a change. Now it will only reflect the latest major update and any bug fixes for that update.